:recycle:[refactor]: 포폴, 팀원모집 공용 ui목적 폴더, 파일구조 변경
