+++ src/app/components/upload/both/FileType.tsx
