+++ src/app/components/upload/both/MiniFileType.tsx
