+++ src/app/components/upload/portfolio/containers/UploadContainerP.tsx
@@ -1,10 +1,11 @@
-import AddThumbnail from '../AddThumbnail';
-import SelectFile from '../SelectFile';
+import { portThumbnail } from 'ui/bothUI';
+import AddThumbnail from '../../both/AddThumbnail';
+import SelectFile from '../SelectFileP';
 
-const UploadContainer = () => {
+const UploadContainerP = () => {
   return (
     <div className="flex flex-col w-full h-[1100px] bg-white">
-      <AddThumbnail />
+      <AddThumbnail thumbnailInfo={portThumbnail} />
       <div className="flex flex-col w-full h-[740px]">
         <div>여긴 컨텐츠들 들어갈 곳</div>
       </div>
@@ -15,4 +16,4 @@ const UploadContainer = () => {
   );
 };
 
-export default UploadContainer;
+export default UploadContainerP;
